+++ addons/gut/utils.gd
@@ -99,8 +99,8 @@ static var InputSender = LazyLoader.new("res://addons/gut/input_sender.gd"):
 static var JunitXmlExport = LazyLoader.new('res://addons/gut/junit_xml_export.gd'):
 	get: return JunitXmlExport.get_loaded()
 	set(val): pass
-static var Logger = LazyLoader.new('res://addons/gut/logger.gd') : # everything should use get_logger
-	get: return Logger.get_loaded()
+static var GutLogger = LazyLoader.new('res://addons/gut/logger.gd') : # everything should use get_logger
+	get: return GutLogger.get_loaded()
 	set(val): pass
 static var MethodMaker = LazyLoader.new('res://addons/gut/method_maker.gd'):
 	get: return MethodMaker.get_loaded()
@@ -159,7 +159,7 @@ static var avail_fonts = ['AnonymousPro', 'CourierPrime', 'LobsterTwo', 'Default
 
 static var version_numbers = VersionNumbers.new(
 	# gut_versrion (source of truth)
-	'9.3.1',
+	'9.4.0',
 	# required_godot_version
 	'4.2.0'
 )
@@ -184,10 +184,10 @@ static var _test_mode = false
 static var _lgr = null
 static func get_logger():
 	if(_test_mode):
-		return Logger.new()
+		return GutLogger.new()
 	else:
 		if(_lgr == null):
-			_lgr = Logger.new()
+			_lgr = GutLogger.new()
 		return _lgr