+++ CHANGES.md
@@ -4,10 +4,28 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 # 9.4.0
 
+## Potentially Breaking Changes
+* The deprecated `wait_frames` and the new `wait_idle_frames`, `wait_physics_frames` now count frames when the `SceneTree.process_frame` and `SceneTree.physics_frame` signals are emitted.  This may cause some of your awaits to wait a frame too long/short.  This approach is more reliable as it occurs before the `_process` or `_physics_process` is called on anything in the tree.  This means that tree order will not matter and all objects will have finished `_process`/`_physics_process` by the time the await ends.
+
+
 ## Features
 * Utilized the adapted Godot tools that generate HTML from code comments, moving some documentation to code comments.  This makes more documentation easily accessible from the editor and cuts down on some duplicate documentation.
+* `wait_idle_frames` added.  This counts frames idle/process frames instead of `_physics_process`.  `wait_frames` has been renamed (deprecated) to `wait_physics_frames`.
+* `wait_while` added.  This waits while a `Callable` returns `true`.
 * New `class_name`s:
   * `GutInputFactory` for `res://addons/gut/gut_input_factory.gd` static class.
+* Signal related methods now accept a reference to a signal as well as an object/signal name:  `get_signal_emit_count`, `assert_connected`, `assert_not_connected`, `assert_signal_emitted`, `assert_signal_not_emitted`, `assert_signal_emitted_with_parameters`, `assert_signal_emit_count`, `get_signal_parameters`.
+
+## Bug Fixes
+* I192 in just under 5 years I moved the two lines of code up some to finally fix this.
+* I666 Using a doubled scene as the value of a property with type `PackedScene` could cause errors.
+
+
+## Deprecations
+* `wait_frames` has been deprecated in lieu of the more specific `wait_physics_frames` and `wait_idle_frames`.
+
+
+
 
 
 # 9.3.1
@@ -37,6 +55,8 @@ A small collection of bug fixes and documentation.  GUT can now generate documen
 * @onegm made their first contribution in https://github.com/bitwes/Gut/pull/668
 
 **Full Changelog**: https://github.com/bitwes/Gut/compare/v9.3.0...9.4.0
+
+
 
 
 # 9.3.0